+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix a bug where the application window isn't in the monitors coordinates
 - Fix a bug where the previous old save is used if there isn't any old save currently
 - Fix a bug where the upload_date is not retrieved from old saves
+- Fix somes bugs where the sort order of new replies are not applied correctly
 
 ## [1.0.0] - 2025-03-23
 - First version of the application with all the essential features.